Move ActionEndpoint and ActionExecutor into backends submodule
