+++ contextshell/CommandInterpreter.py
@@ -1,6 +1,6 @@
 from contextshell.Command import Command
 from contextshell.NodePath import NodePath
-from contextshell.ActionExecutor import ActionExecutor, parse_argument_tree
+from contextshell.backends.ActionExecutor import ActionExecutor, parse_argument_tree
 
 
 class CommandInterpreter:
