+++ tests/unit/VirtualTreeTests.py
@@ -1,6 +1,6 @@
 import unittest
 
-from contextshell.ActionExecutor import ActionExecutor, ActionArgsPack, OrderedDict, pack_argument_tree
+from contextshell.backends.ActionExecutor import ActionExecutor, ActionArgsPack, OrderedDict, pack_argument_tree
 from contextshell.NodePath import NodePath
